+++ Tensile/KernelWriterSource.py
@@ -832,22 +832,22 @@ class KernelWriterSource(KernelWriter):
     if self.language == "HIP":
       #s += "  hipLaunchParm lp," + self.endLine
       globalStr = ""
-    restrictStr = "restrict"
-    if self.language == "HIP":
-      restrictStr = "__restrict__"
     ptrStr = kernel["ProblemType"]["DestDataType"].toDevice(self.language) \
         if not kernel["_GlobalAccumulation"] else "float"
-    s += "  " + globalStr + ptrStr + " *D,"
+    isStridedBuffer = kernel["ProblemType"]["StridedBatched"] or kernel["_GlobalAccumulation"]
+    ptrStr  += ("" if isStridedBuffer else "*")
+    batchStr = ("" if isStridedBuffer else "Batch")
+    s += "  " + globalStr + ptrStr + " *"+ batchStr + "D,"
     s += self.endLine
-    s += "  " + globalStr + ptrStr \
-        + " const * " + restrictStr + " C,"
+    s += "  " + globalStr + ptrStr + " const * " + batchStr + "C,"
     s += self.endLine
-    ptrStr = kernel["ProblemType"]["DataType"].toDevice(self.language)
-    s += "  " + globalStr + ptrStr \
-        + " const * " + restrictStr + " A,"
+
+    ptrStr   = kernel["ProblemType"]["DataType"].toDevice(self.language)
+    ptrStr  += ("" if kernel["ProblemType"]["StridedBatched"] else "*")
+    batchStr = ("" if kernel["ProblemType"]["StridedBatched"] else "Batch")
+    s += "  " + globalStr + ptrStr + " const * " + batchStr + "A,"
     s += self.endLine
-    s += "  " + globalStr + ptrStr \
-        + " const * " + restrictStr + " B"
+    s += "  " + globalStr + ptrStr + " const * " + batchStr + "B"
 
     # alpha & beta
     s += "," + self.endLine + "  " \
@@ -907,7 +907,13 @@ class KernelWriterSource(KernelWriter):
     # kernel["PersistentKernel"]:
     s += "," + self.endLine + "  unsigned int problemNumGroupTiles0"
     s += "," + self.endLine + "  unsigned int problemNumGroupTiles1"
-    s += "," + self.endLine + "  unsigned int magicNumberProblemNumGroupTiles0"
+
+    # offset
+    s += "," + self.endLine + "  unsigned int offsetD"
+    s += "," + self.endLine + "  unsigned int offsetC"
+    s += "," + self.endLine + "  unsigned int offsetA"
+    s += "," + self.endLine + "  unsigned int offsetB"
+
     s += " )"
     return s
 
@@ -928,6 +934,7 @@ class KernelWriterSource(KernelWriter):
     s = ""
     s += " {" + self.endLine
     return s
+
 
   ##############################################################################
   # Allocate Resources
@@ -1031,6 +1038,26 @@ class KernelWriterSource(KernelWriter):
         % (self.sharedDeclStr, self.endLine )
 
 
+    ####################################
+    # apply general batch
+    if not kernel["ProblemType"]["StridedBatched"]:
+      kStr += self.endLine
+      kStr += "  unsigned int wg = " + self.getGroupIdStr + "(2);" + self.endLine
+      if not kernel["_GlobalAccumulation"]:
+        kStr += "  DEST_DATA_TYPE      * D = BatchD[wg];" + self.endLine
+        kStr += "  DEST_DATA_TYPE const* C = BatchC[wg];" + self.endLine
+      kStr += "  DATA_TYPE      const* A = BatchA[wg];" + self.endLine
+      kStr += "  DATA_TYPE      const* B = BatchB[wg];" + self.endLine
+
+    ####################################
+    # apply offset
+    kStr += self.endLine
+    if not kernel["_GlobalAccumulation"]:
+      kStr += "  D = D + offsetD;" + self.endLine
+      kStr += "  C = C + offsetC;" + self.endLine
+    kStr += "  A = A + offsetA;" + self.endLine
+    kStr += "  B = B + offsetB;" + self.endLine
+
     if 0:
       # in some cases we know the pad values at compile time and could hard-code here.  Not enabled.
       for tc in ('A', 'B'):
@@ -1050,8 +1077,9 @@ class KernelWriterSource(KernelWriter):
     self.magicNonSumChars = kernel["PackedC0IdxChars"][:-1] + kernel["PackedC1IdxChars"][:-1]
 
     if kernel["MagicDivAlg"] == 2:
-      kStr += "typedef struct MagicStruct {unsigned M; int a; int s;} MagicStruct;" + self.endLine
-      kStr += "const unsigned MAGIC_STRUCT_A = 0x80000000; // for extracting a-bit from shift kernarg" + self.endLine
+      kStr += self.endLine
+      kStr += "  typedef struct MagicStruct {unsigned M; int a; int s;} MagicStruct;" + self.endLine
+      kStr += "  const unsigned MAGIC_STRUCT_A = 0x80000000; // for extracting a-bit from shift kernarg" + self.endLine
       kStr += "#define MAGIC_DIV2(dividend, magic) (((((uint64_t)(dividend) * magic.M) >> 32) + dividend*magic.a) >> magic.s)%s" % self.endLine
 
       sumParms=[(idxChar, "magicStruct%s"%idxChar, "NumIter%s"%idxChar) for idxChar in self.magicSumChars]
@@ -1120,16 +1148,20 @@ class KernelWriterSource(KernelWriter):
         kStr += "  %s  = serialWgIter %% numWGIJ;%s" % ( wgIJSerial, self.endLine)
         kStr += "  %s  = %s %% problemNumGroupTiles0;%s" % ( wg0, wgIJSerial, self.endLine)
         kStr += "  %s  = %s / problemNumGroupTiles0;%s" % ( wg1, wgIJSerial, self.endLine)
+        if not kernel["ProblemType"]["StridedBatched"]:
+          if not kernel["_GlobalAccumulation"]:
+            kStr += "  D = BatchD[wgKSerial] + offsetD;%s" % self.endLine
+            kStr += "  C = BatchC[wgKSerial] + offsetC;%s" % self.endLine
+          kStr += "  A = BatchA[wgKSerial] + offsetA;%s" % self.endLine
+          kStr += "  B = BatchB[wgKSerial] + offsetB;%s" % self.endLine
       else:
         # compare serialWgIter against problem groups
         if kernel["GlobalSplitU"] > 1:
           kStr += "  if (serialWgIter >= numWGIJ * GLOBAL_SPLITU) break; // persistent loop" + self.endLine
         else:
           kStr += "  if (serialWgIter >= numWGIJ) break; // persistent loop" + self.endLine
-        kStr += "  %s  = serialWgIter %% problemNumGroupTiles0;%s" \
-            % ( wg0, self.endLine)
-        kStr += "  %s  = serialWgIter / problemNumGroupTiles0;%s" \
-            % ( wg1, self.endLine)
+        kStr += "  %s  = serialWgIter %% problemNumGroupTiles0;%s" % ( wg0, self.endLine)
+        kStr += "  %s  = serialWgIter / problemNumGroupTiles0;%s" % ( wg1, self.endLine)
     else:
       # optionally transpose work-group grid
       kStr += "  unsigned int %s = %s(%u);%s" \
@@ -1276,6 +1308,7 @@ class KernelWriterSource(KernelWriter):
           index2 = nonTileFreeIndices[j]
           kStr += " / size" + self.indexChars[index2]
         kStr += " ) % size" + self.indexChars[index] + ";" + self.endLine
+
     return kStr
 
   ##############################################################################
@@ -1474,7 +1507,10 @@ class KernelWriterSource(KernelWriter):
                           (sPara if tP["tlu"] else sPerp) )
                   else:
                     # just a non-vector group index
-                    kStr += "wg" + self.indexChars[index]
+                    if kernel["ProblemType"]["StridedBatched"]:
+                      kStr += "wg" + self.indexChars[index]
+                    else:
+                      kStr += "0"
               else: # summation index
                 if index == kernel["ProblemType"]["IndexUnroll"]:
                   kStr += "(globalReadOffset%s%s_%u_%u)" \
@@ -1524,6 +1560,7 @@ class KernelWriterSource(KernelWriter):
   ##############################################################################
   def graAddresses(self, kernel, tP):
     kStr = ""
+
     for perp in range(0, tP["nrp"]):
       for sPerp in range(0, tP["nrpv"]):
         for para in range(0, tP["nrc"]):
@@ -2940,17 +2977,18 @@ class KernelWriterSource(KernelWriter):
             % (self.indexChars[index1], self.tileChar1, self.tileChar0, self.endLine)
 
     for i in range(0, kernel["ProblemType"]["NumIndicesC"]):
-        kStr += "  unsigned int globalC%s = " % self.indexChars[i]
-        if i == index0 and len(kernel["PackedC0IndicesX"])==1:
-          kStr += "flattenedGlobalC0;"
-        elif i == index1 and len(kernel["PackedC1IndicesX"])==1:
-          kStr += "flattenedGlobalC1;"
-        elif isPackedIndex(kernel,i):
-          kStr += "0; // will be set below"
-        else:
-          #kStr += "printf(\"pre: serial:%%u wg0:%%u wg1:%%u globalC0I:%%u globalC1J:%%u\\n\", serial, wg0I, wg1J, globalC0I, globalC1J);%s" % (self.endLine)
-          kStr += "(wg%s);" % (self.indexChars[i])
-        kStr += "%s" % self.endLine
+      kStr += "  unsigned int globalC%s = " % self.indexChars[i]
+      if i == index0 and len(kernel["PackedC0IndicesX"]) == 1:
+        kStr += "flattenedGlobalC0;"
+      elif i == index1 and len(kernel["PackedC1IndicesX"]) == 1:
+        kStr += "flattenedGlobalC1;"
+      elif isPackedIndex(kernel,i):
+        kStr += "0; // will be set below"
+      elif kernel["ProblemType"]["StridedBatched"] or kernel["_GlobalAccumulation"]:
+        kStr += "(wg%s);" % (self.indexChars[i])
+      else:
+        kStr += "0;"
+      kStr += "%s" % self.endLine
 
     if kernel["_GlobalAccumulation"] == 'MultipleBuffer':
       indexChar = self.indexChars[0]