+++ Tensile/Source/client/source/ConvolutionProblem.cpp
@@ -436,7 +436,8 @@ namespace Tensile
                             activationDims.begin(),
                             activationDims.end(),
                             activationStri.begin(),
-                            activationStri.end());
+                            activationStri.end(),
+                            problem.a().offset());
         return rv;
     }
 
@@ -468,7 +469,8 @@ namespace Tensile
         default:
             throw std::runtime_error("unknown formatD");
         };
-        TensorDescriptor rv(problem.d().dataType(), outputDims.begin(), outputDims.end());
+        TensorDescriptor rv(
+            problem.d().dataType(), outputDims.begin(), outputDims.end(), problem.d().offset());
         return rv;
     }
 
@@ -496,7 +498,8 @@ namespace Tensile
         default:
             throw std::runtime_error("unknown formatB");
         };
-        TensorDescriptor rv(problem.b().dataType(), filterDims.begin(), filterDims.end());
+        TensorDescriptor rv(
+            problem.b().dataType(), filterDims.begin(), filterDims.end(), problem.b().offset());
         return rv;
     }